De-fragment and Display data received from the device
Currently relies on calls from Flask to trigger commands to be sent, and data to be acessed. This should be done as asynchronous events, but this would require the device to have a main thread and the processing thread.

Needs timeouts to account for lost packets unless data is sent from device with MQTT-Quality of Service = 2.

Signed-off-by: unknown <sb@Stephen-PC.internal.audioanalytic.com>